Analyzing Customer Behavior and Risk Factors
This project focuses on analyzing customer behavior and identifying risk factors associated with loan defaults. The objective is to help banks make informed decisions by pinpointing the types of loans that are most likely to default and identifying the factors of customers who are at a higher risk of defaulting on their loans.